aaaaaaaa

+++ jpa/src/main/java/br/edu/ufersa/CatCaffe/models/entities/Funcionario.java
@@ -9,6 +9,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -23,7 +24,7 @@ public class Funcionario extends Usuario{
     private String nome;
 
     @Column(nullable = false)
-    private int salario;
+    private BigDecimal salario;
 
     @Column(nullable = false,length = 50)
     private String cargo;
